+++ src/main/php/ioc/Binder.php
@@ -21,6 +21,48 @@ use stubbles\lang\Properties;
 class Binder
 {
     /**
+     * switch whether setter injection is enabled
+     *
+     * @type  bool
+     * @since  5.1.0
+     * @deprecated  since 5.1.0, don't use setter injection, will be removed with 6.0.0
+     */
+    private static $setterInjectionEnabled = false;
+
+    /**
+     * enable setter injection
+     *
+     * @since  5.1.0
+     * @deprecated  since 5.1.0, don't use setter injection, will be removed with 6.0.0
+     */
+    public static function enableSetterInjection()
+    {
+        self::$setterInjectionEnabled = true;
+    }
+
+    /**
+     * disable setter injection
+     *
+     * @since  5.1.0
+     * @deprecated  since 5.1.0, will be removed with 6.0.0
+     */
+    public static function disableSetterInjection()
+    {
+        self::$setterInjectionEnabled = false;
+    }
+    /**
+     * checks whether setter injection is enabled
+     *
+     * @return  bool
+     * @since  5.1.0
+     * @deprecated  since 5.1.0, don't use setter injection, will be removed with 6.0.0
+     */
+    public static function isSetterInjectionEnabled()
+    {
+        return self::$setterInjectionEnabled;
+    }
+
+    /**
      * index for faster access to bindings
      *
      * @type  \stubbles\ioc\binding\BindingIndex