+++ src/test/helper/ioc/AnotherQuestion.php
@@ -25,7 +25,7 @@ class AnotherQuestion
      * @Inject
      * @Named('answer')
      */
-    public function setAnswer(Answer $answer)
+    public function __construct(Answer $answer)
     {
         $this->answer = $answer;
     }